+++ CHANGELOG.md
@@ -29,3 +29,15 @@
 
 - Some pathing and code optimizations.
 - Fixed a lot of pathing issues.
+
+## 0.1.0
+
+- Added soft dependancy on Coroner 
+    - For custom cause of death for gargoyle.
+    - Allows access to more causes of deaths in prior death taunts.
+- Mod now tracks prior round deaths
+    -  Adding 80 more taunts relating to cause of death. (18 from vanilla, 62 from Coroner.)
+    - Gargoyle will now taunt it's target with how they died the last round.
+- Added 1 aggro and 2 general taunts.
+- A few tweaks to the settings.
+- Default Idle config reduced from 30 to 20. That way Gargoyles should follow their target a bit more closely.